Remove last_offer_successful logic
So the rationale behind the previous implementation was that I
mistakenly assumed it was wrong to assign nullptr to new_pc (now renamed
tmp_pc) inside its callback definition. But the object tmp_pc points to
was already passed to task->m_pc and, by replacing the "this" pointer
(which I think was wrong in the first place) by tmp_pc, which is passed
through the capture clause, I make sure tmp_pc gets null while the
PeerConnection object is unaffected and still pointed to by task->m_pc.